refactor: create smart and dumb components for better code management
