+++ src/features/products/components/ProductDetails/ErrorDisplay.jsx
@@ -1,0 +1,35 @@
+import React from "react";
+import { Link } from "react-router-dom";
+
+const ErrorDisplay = ({ error }) => {
+  return (
+    <div className="min-h-screen flex items-center justify-center bg-gray-50">
+      <div className="bg-white p-8 rounded-lg shadow-md max-w-lg mx-auto text-center">
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          className="h-16 w-16 text-red-500 mx-auto mb-4"
+          fill="none"
+          viewBox="0 0 24 24"
+          stroke="currentColor"
+        >
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth={2}
+            d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
+          />
+        </svg>
+        <h2 className="text-2xl font-bold text-gray-800 mb-2">Error</h2>
+        <p className="text-gray-600 mb-6">{error}</p>
+        <Link
+          to="/products"
+          className="inline-block bg-indigo-600 text-white px-6 py-3 rounded-lg font-medium hover:bg-indigo-700 transition-colors"
+        >
+          Return to Products
+        </Link>
+      </div>
+    </div>
+  );
+};
+
+export default ErrorDisplay;
