+++ src/features/products/components/ProductDetails/LoadingSpinner.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const LoadingSpinner = () => {
+  return (
+    <div className="min-h-screen flex items-center justify-center bg-gray-50">
+      <div className="w-16 h-16 border-4 border-indigo-200 border-t-indigo-600 rounded-full animate-spin"></div>
+    </div>
+  );
+};
+
+export default LoadingSpinner;
