+++ src/features/products/components/ProductDetails/NotFoundDisplay.jsx
@@ -1,0 +1,40 @@
+import React from "react";
+import { Link } from "react-router-dom";
+
+const NotFoundDisplay = () => {
+  return (
+    <div className="min-h-screen flex items-center justify-center bg-gray-50">
+      <div className="text-center bg-white p-8 rounded-lg shadow-md max-w-lg mx-auto">
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          className="h-16 w-16 text-gray-400 mx-auto mb-4"
+          fill="none"
+          viewBox="0 0 24 24"
+          stroke="currentColor"
+        >
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth={2}
+            d="M9.172 16.172a4 4 0 015.656 0M9 10h.01M15 10h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"
+          />
+        </svg>
+        <h2 className="text-2xl font-bold text-gray-800 mb-2">
+          Product not found
+        </h2>
+        <p className="text-gray-600 mb-6">
+          The product you're looking for doesn't exist or has been removed.
+        </p>
+        <Link
+          to="/products"
+          className="inline-block bg-indigo-600 text-white px-6 py-3 rounded-lg font-medium hover:bg-indigo-700 transition-colors"
+        >
+          Browse Products
+        </Link>
+      </div>
+    </div>
+  );
+};
+
+
+export default NotFoundDisplay;
