+++ src/features/products/components/ProductDetails/ProductReviews.jsx
@@ -1,0 +1,36 @@
+// ProductReviews.jsx
+import React from "react";
+
+const ProductReviews = () => {
+  return (
+    <div className="mb-10">
+      <div className="flex items-center justify-between mb-6 pb-2 border-b border-gray-200">
+        <h2 className="text-2xl font-bold text-gray-900">Customer Reviews</h2>
+        <button className="inline-flex items-center px-4 py-2 border border-transparent text-sm font-medium rounded-md shadow-sm text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500">
+          Write a Review
+        </button>
+      </div>
+      <div className="text-center py-12">
+        <svg
+          className="mx-auto h-12 w-12 text-gray-400"
+          fill="none"
+          viewBox="0 0 24 24"
+          stroke="currentColor"
+        >
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth={2}
+            d="M7 8h10M7 12h4m1 8l-4-4H5a2 2 0 01-2-2V6a2 2 0 012-2h14a2 2 0 012 2v8a2 2 0 01-2 2h-3l-4 4z"
+          />
+        </svg>
+        <h3 className="mt-2 text-sm font-medium text-gray-900">No reviews yet</h3>
+        <p className="mt-1 text-sm text-gray-500">
+          Be the first to review this product.
+        </p>
+      </div>
+    </div>
+  );
+};
+
+export default ProductReviews;
